refactor: move AiAgentService to dedicated directory with documentation
