+++ packages/backend/src/ee/scheduler/SchedulerWorker.ts
@@ -12,7 +12,7 @@ import { tryJobOrTimeout } from '../../scheduler/SchedulerJobTimeout';
 import { SchedulerTaskArguments } from '../../scheduler/SchedulerTask';
 import { SchedulerWorker } from '../../scheduler/SchedulerWorker';
 import { TypedEETaskList } from '../../scheduler/types';
-import { AiAgentService } from '../services/AiAgentService';
+import { AiAgentService } from '../services/AiAgentService/AiAgentService';
 import type { EmbedService } from '../services/EmbedService/EmbedService';
 
 const AI_AGENT_EVAL_TIMEOUT_MS = 10 * 60 * 1000; // 10 minutes in milliseconds
